mdcode: rename deploy.ts to deploy_bigquery.ts for symmetry
deploy.ts is entirely BigQuery-specific (deployBigQuery,
bigQueryGraphTargets), so name it by destination to match the
knowledge_catalog pair. The semantic/ layer is now symmetric:

  bigquery.ts / deploy_bigquery.ts             (BigQuery: emit / publish)
  knowledge_catalog.ts / deploy_knowledge_catalog.ts   (KC: emit / publish)

Pure rename plus import/comment updates; no behavior change.
